--- a/pkg/storage/tsdb/ref_cache.go
+++ b/pkg/storage/tsdb/ref_cache.go
@@ -27,20 +27,20 @@ const (
 type RefCache struct {
 	// The cache is split into stripes, each one with a dedicated lock, in
 	// order to reduce lock contention.
-	stripes [numRefCacheStripes]*refCacheStripe
+	stripes [numRefCacheStripes]refCacheStripe
 }
 
 // refCacheStripe holds a subset of the series references for a single tenant.
 type refCacheStripe struct {
 	refsMu sync.Mutex
-	refs   map[model.Fingerprint][]*refCacheEntry
+	refs   map[model.Fingerprint][]refCacheEntry
 }
 
 // refCacheEntry holds a single series reference.
 type refCacheEntry struct {
 	lbs       labels.Labels
 	ref       uint64
-	touchedAt time.Time
+	touchedAt int64 // Unix nano time.
 }
 
 // NewRefCache makes a new RefCache.
@@ -49,9 +49,7 @@ func NewRefCache() *RefCache {
 
 	// Stripes are pre-allocated so that we only read on them and no lock is required.
 	for i := 0; i < numRefCacheStripes; i++ {
-		c.stripes[i] = &refCacheStripe{
-			refs: map[model.Fingerprint][]*refCacheEntry{},
-		}
+		c.stripes[i].refs = map[model.Fingerprint][]refCacheEntry{}
 	}
 
 	return c
@@ -61,7 +59,7 @@ func NewRefCache() *RefCache {
 // is NOT retained.
 func (c *RefCache) Ref(now time.Time, series labels.Labels) (uint64, bool) {
 	fp := client.Fingerprint(series)
-	stripeID := uint8(util.HashFP(fp) % numRefCacheStripes)
+	stripeID := util.HashFP(fp) % numRefCacheStripes
 
 	return c.stripes[stripeID].ref(now, series, fp)
 }
@@ -69,7 +67,7 @@ func (c *RefCache) Ref(now time.Time, series labels.Labels) (uint64, bool) {
 // SetRef sets/updates the cached series reference. The input labels set IS retained.
 func (c *RefCache) SetRef(now time.Time, series labels.Labels, ref uint64) {
 	fp := client.Fingerprint(series)
-	stripeID := uint8(util.HashFP(fp) % numRefCacheStripes)
+	stripeID := util.HashFP(fp) % numRefCacheStripes
 
 	c.stripes[stripeID].setRef(now, series, fp, ref)
 }
@@ -91,13 +89,12 @@ func (s *refCacheStripe) ref(now time.Time, series labels.Labels, fp model.Finge
 		return 0, false
 	}
 
-	for _, entry := range entries {
+	for ix, entry := range entries {
 		if labels.Equal(entry.lbs, series) {
-			// Get the reference and touch the timestamp before releasing the lock
-			ref := entry.ref
-			entry.touchedAt = now
-
-			return ref, true
+			// Touch the timestamp before releasing the lock
+			//entry.touchedAt = now.UnixNano()
+			entries[ix].touchedAt = now.UnixNano()
+			return entry.ref, true
 		}
 	}
 
@@ -109,18 +106,19 @@ func (s *refCacheStripe) setRef(now time.Time, series labels.Labels, fp model.Fi
 	defer s.refsMu.Unlock()
 
 	// Check if already exists within the entries.
-	for _, entry := range s.refs[fp] {
+	for ix, entry := range s.refs[fp] {
 		if !labels.Equal(entry.lbs, series) {
 			continue
 		}
 
 		entry.ref = ref
-		entry.touchedAt = now
+		entry.touchedAt = now.UnixNano()
+		s.refs[fp][ix] = entry
 		return
 	}
 
 	// The entry doesn't exist, so we have to add a new one.
-	s.refs[fp] = append(s.refs[fp], &refCacheEntry{lbs: series, ref: ref, touchedAt: now})
+	s.refs[fp] = append(s.refs[fp], refCacheEntry{lbs: series, ref: ref, touchedAt: now.UnixNano()})
 }
 
 func (s *refCacheStripe) purge(keepUntil time.Time) {
@@ -131,7 +129,7 @@ func (s *refCacheStripe) purge(keepUntil time.Time) {
 		// Since we do expect very few fingerprint collisions, we
 		// have an optimized implementation for the common case.
 		if len(entries) == 1 {
-			if entries[0].touchedAt.Before(keepUntil) {
+			if entries[0].touchedAt < keepUntil.UnixNano() {
 				delete(s.refs, fp)
 			}
 
@@ -141,7 +139,7 @@ func (s *refCacheStripe) purge(keepUntil time.Time) {
 		// We have more entries, which means there's a collision,
 		// so we have to iterate over the entries.
 		for i := 0; i < len(entries); {
-			if entries[i].touchedAt.Before(keepUntil) {
+			if entries[i].touchedAt < keepUntil.UnixNano() {
 				entries = append(entries[:i], entries[i+1:]...)
 			} else {
 				i++