**What this PR does**: this PR does some changes to `RefCache` to improve its throughput in new `RefCacheConcurrency` benchmark. Unfortunately it also makes individual `Ref` and `SetRef` calls slower. Sending for discussion, not necessarily merging.

```
name                           old time/op  new time/op  delta
RefCacheConcurrency_1m_50-4     1.29s ± 4%   1.10s ± 5%  -14.84%  (p=0.000 n=9+10)
RefCacheConcurrency_1m_250-4    2.35s ± 3%   1.85s ± 6%  -21.46%  (p=0.000 n=10+10)
RefCacheConcurrency_1m_1000-4   10.6s ± 2%    7.8s ± 3%  -27.00%  (p=0.000 n=10+10)
RefCache_SetRef-4               414µs ± 2%   651µs ± 2%  +57.43%  (p=0.000 n=10+10)
RefCache_Ref-4                  815µs ± 2%   894µs ± 3%   +9.69%  (p=0.000 n=10+10)
RefCache_purge-4                102ms ± 2%    54ms ± 2%  -47.13%  (p=0.000 n=10+10)
```