@pstibrany 2020-05-29
> * Would you mind measuring mem allocations diff too, please?

Benchmark with allocations:

```
$ benchstat before_mem.txt after_mem.txt 
name                           old time/op    new time/op    delta
RefCacheConcurrency_1m_50-4       1.28s ± 9%     1.05s ±13%   -17.55%  (p=0.000 n=10+9)
RefCacheConcurrency_1m_250-4      2.24s ± 7%     1.67s ± 2%   -25.52%  (p=0.000 n=9+10)
RefCacheConcurrency_1m_1000-4     10.7s ±15%      7.2s ± 2%   -32.15%  (p=0.000 n=9+10)
RefCache_SetRef-4                 393µs ± 3%     657µs ±12%   +67.16%  (p=0.000 n=10+10)
RefCache_Ref-4                    783µs ± 2%     826µs ± 4%    +5.42%  (p=0.000 n=10+10)
RefCache_purge-4                 95.7ms ± 3%    49.9ms ± 2%   -47.87%  (p=0.000 n=9+10)

name                           old alloc/op   new alloc/op   delta
RefCacheConcurrency_1m_50-4       235MB ± 0%     214MB ± 0%    -8.85%  (p=0.000 n=10+8)
RefCacheConcurrency_1m_250-4      235MB ± 0%     214MB ± 0%    -8.87%  (p=0.000 n=10+10)
RefCacheConcurrency_1m_1000-4     235MB ± 0%     214MB ± 0%    -8.87%  (p=0.000 n=10+10)
RefCache_SetRef-4                 0.00B          0.00B           ~     (all equal)
RefCache_Ref-4                    0.00B          0.00B           ~     (all equal)
RefCache_purge-4                   768B ± 0%     1706B ± 0%  +122.14%  (p=0.000 n=10+8)

name                           old allocs/op  new allocs/op  delta
RefCacheConcurrency_1m_50-4       2.03M ± 0%     1.02M ± 0%   -49.63%  (p=0.000 n=10+8)
RefCacheConcurrency_1m_250-4      2.03M ± 0%     1.02M ± 0%   -49.64%  (p=0.000 n=10+10)
RefCacheConcurrency_1m_1000-4     2.03M ± 0%     1.02M ± 0%   -49.63%  (p=0.000 n=10+10)
RefCache_SetRef-4                  0.00           0.00           ~     (all equal)
RefCache_Ref-4                     0.00           0.00           ~     (all equal)
RefCache_purge-4                   21.0 ± 0%      21.0 ± 0%      ~     (all equal)
```

> * Why is the single get/set slower? What makes it slower?

`SetRef` is slower because locking unlocked `RWMutex` is bit slower than `Mutex`. `Ref` is slower because of `RWMutex` and atomic update of timestamp.