@pracucci 2020-05-29
> SetRef is slower because locking unlocked RWMutex is bit slower than Mutex. Ref is slower because of RWMutex and atomic update of timestamp.

That's fine. Supporting higher concurrency here is better.